// ui/src/utils/types/router.ts
export enum RouteName {
  LOGIN = 'LOGIN',
  CHORUS_HOME = 'CHORUS_HOME',
  CHORUS_REPLICATION = 'CHORUS_REPLICATION',
  CHORUS_ADD_REPLICATION = 'CHORUS_ADD_REPLICATION',
  CHORUS_STORAGES = 'CHORUS_STORAGES',
  CHORUS_STORAGE_DETAILS = 'CHORUS_STORAGE_DETAILS',
  CHORUS_MONITORING = 'CHORUS_MONITORING',
}